Potential fix for code scanning alert no. 55: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using a larger integer type to avoid overflow. This can be achieved by casting one of the operands to unsigned long before performing the multiplication. This way, the multiplication will be done using unsigned long arithmetic, which has a larger range and can accommodate larger results without overflow.

Specifically, we will cast (nbstates + 1) to unsigned long before performing the multiplication with (nbatoms + 1).

+++ libxml2-2.9.9/xmlregexp.c
@@ -540,8 +540,8 @@ xmlRegEpxFromParse(xmlRegParserCtxtPtr ctxt) {
 #ifdef DEBUG_COMPACTION
 	printf("Final: %d atoms\n", nbatoms);
 #endif
-	transitions = (int *) xmlMalloc(((unsigned long)(nbstates + 1)) *
-	                                (nbatoms + 1) * sizeof(int));
+	transitions = (int *) xmlMalloc((unsigned long)(nbstates + 1) *
+	                                (unsigned long)(nbatoms + 1) * sizeof(int));
 	if (transitions == NULL) {
 	    xmlFree(stateRemap);
 	    xmlFree(stringRemap);
@@ -549,7 +549,7 @@ xmlRegEpxFromParse(xmlRegParserCtxtPtr ctxt) {
 	    xmlFree(ret);
 	    return(NULL);
 	}
-	memset(transitions, 0, (nbstates + 1) * (nbatoms + 1) * sizeof(int));
+	memset(transitions, 0, (unsigned long)(nbstates + 1) * (unsigned long)(nbatoms + 1) * sizeof(int));
 
 	/*
 	 * Allocate the transition table. The first entry for each
